services/hrms-tools/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# LLM Providers
openai==1.10.0
anthropic==0.18.0
google-generativeai==0.3.2

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0

# HTTP and utilities
httpx==0.26.0
python-multipart==0.0.6

# Logging and monitoring
python-json-logger==2.0.7
